requirements.txt
pandas 
mlflow==2.2.2
notebook
numpy
scikit-learn
matplotlib
python-box
pyYAML
tqdm
joblib
types-PyYAML
Flask
Flask-Cors
dill
imblearn
xgboost
ipykernel 
seaborn
matplotlib
langchain_groq
langchain_community
faiss-cpu
python-dotenv
pypdf
langchain_huggingface
markdown-it-py==3.0.0
mdurl==0.1.2
pygments==2.19.1
rich==13.9.4
streamlit==1.42.2
groq
-e .